+++ extension/datadogfleetautomationextension/http_test.go
@@ -545,117 +545,3 @@ func TestHandleMetadata(t *testing.T) {
 		})
 	}
 }
-
-func TestMakeGetRequest(t *testing.T) {
-	tests := []struct {
-		name           string
-		serverResponse string
-		serverStatus   int
-		expectedError  string
-		expectedBody   string
-		url            string
-	}{
-		{
-			name:           "Successful response",
-			serverResponse: `{"status": "ok"}`,
-			serverStatus:   http.StatusOK,
-			expectedError:  "",
-			expectedBody:   `{"status": "ok"}`,
-		},
-		{
-			name:           "Server error",
-			serverResponse: `Internal Server Error`,
-			serverStatus:   http.StatusInternalServerError,
-			expectedError:  "",
-			expectedBody:   `Internal Server Error`,
-		},
-		{
-			name:           "Invalid URL",
-			serverResponse: ``,
-			serverStatus:   http.StatusOK,
-			expectedError:  "missing protocol scheme",
-			expectedBody:   ``,
-			url:            "://invalid-url",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Create a test server
-			server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
-				w.WriteHeader(tt.serverStatus)
-				_, err := w.Write([]byte(tt.serverResponse))
-				assert.NoError(t, err)
-			}))
-			defer server.Close()
-
-			// Determine the URL to use for the test
-			if tt.url == "" {
-				tt.url = server.URL
-			}
-
-			// Call makeGetRequest
-			resp, err := makeGetRequest(tt.url)
-			defer func() {
-				if resp != nil {
-					_, _ = resp.Body.Read(nil)
-					_ = resp.Body.Close()
-				}
-			}()
-			// Verify the result
-			if tt.expectedError == "" {
-				assert.NoError(t, err)
-				assert.NotNil(t, resp)
-				body := make([]byte, len(tt.expectedBody))
-				_, err = resp.Body.Read(body)
-				assert.ErrorContains(t, err, "EOF")
-				assert.Equal(t, tt.expectedBody, string(body))
-			} else {
-				assert.Error(t, err)
-				assert.Contains(t, err.Error(), tt.expectedError)
-				assert.Nil(t, resp)
-			}
-		})
-	}
-}
-
-func TestIsLocalAddress(t *testing.T) {
-	tests := []struct {
-		name     string
-		hostname string
-		expected bool
-	}{
-		{
-			name:     "Localhost",
-			hostname: "localhost",
-			expected: true,
-		},
-		{
-			name:     "IPv4 Loopback",
-			hostname: "127.0.0.1",
-			expected: true,
-		},
-		{
-			name:     "IPv6 Loopback",
-			hostname: "::1",
-			expected: true,
-		},
-		{
-			name:     "Non-local address",
-			hostname: "192.168.1.1",
-			expected: false,
-		},
-		{
-			name:     "Invalid IP address",
-			hostname: "invalid-ip",
-			expected: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := isLocalAddress(tt.hostname)
-			assert.Equal(t, tt.expected, result)
-		})
-	}
-}
